+++ WorldCities/WorldCities/src/app/cities/cities.component.ts
@@ -1,6 +1,9 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { environment } from './../../envrionments/environment';
+
+import { MatTable, MatTableDataSource } from '@angular/material/table';
+import { MatPaginator } from '@angular/material/paginator';
 
 import { City } from './city';
 
@@ -13,14 +16,17 @@ export class CitiesComponent implements OnInit {
 
   public displayedColumns: string[] = ['id', 'name', 'lat', 'lon'];
 
-  public cities!: City[];
+  public cities!: MatTableDataSource<City>;
+
+  @ViewChild(MatPaginator) paginator!: MatPaginator;
 
   constructor(private http: HttpClient) { }
 
   ngOnInit() {
     this.http.get<City[]>(environment.baseURL + 'api/Cities')
       .subscribe(result => {
-        this.cities = result;
+        this.cities = new MatTableDataSource<City>(result);
+        this.cities.paginator = this.paginator;
       }, error => console.error(error));
   }
 }